Extract ActiveFedora::InitializingConnection to lazy initialize the repository
