+++ lib/active_fedora/fedora.rb
@@ -18,6 +18,8 @@ module ActiveFedora
 
     def initialize(config)
       @config = config
+
+      validate_options
     end
 
     def host
@@ -42,7 +44,6 @@ module ActiveFedora
 
     def connection
       @connection ||= begin
-        init_base_path
         build_connection
       end
     end
@@ -51,28 +52,16 @@ module ActiveFedora
       @clean_connection ||= CleanConnection.new(connection)
     end
 
+    def caching_connection(options = {})
+      CachingConnection.new(authorized_connection, options)
+    end
+
     def ldp_resource_service
       @service ||= LdpResourceService.new(connection)
     end
 
     SLASH = '/'.freeze
     BLANK = ''.freeze
-
-    # Call this to create a Container Resource to act as the base path for this connection
-    def init_base_path
-      return if @initialized
-      connection = build_connection
-
-      connection.head(root_resource_path)
-      ActiveFedora::Base.logger.info "Attempted to init base path `#{root_resource_path}`, but it already exists" if ActiveFedora::Base.logger
-      @initialized = true
-      false
-    rescue Ldp::NotFound
-      unless host.downcase.end_with?("/rest")
-        ActiveFedora::Base.logger.warn "Fedora URL (#{host}) does not end with /rest. This could be a problem. Check your fedora.yml config"
-      end
-      @initialized = connection.put(root_resource_path, BLANK).success?
-    end
 
     # Remove a leading slash from the base_path
     def root_resource_path
@@ -83,8 +72,8 @@ module ActiveFedora
       # The InboundRelationConnection does provide more data, useful for
       # things like ldp:IndirectContainers, but it's imposes a significant
       # performance penalty on every request
-      #   @connection ||= InboundRelationConnection.new(CachingConnection.new(authorized_connection))
-      CachingConnection.new(authorized_connection, omit_ldpr_interaction_model: true)
+      #   @connection ||= InboundRelationConnection.new(caching_connection(omit_ldpr_interaction_model: true))
+      InitializingConnection.new(caching_connection(omit_ldpr_interaction_model: true), root_resource_path)
     end
 
     def authorized_connection
@@ -93,6 +82,12 @@ module ActiveFedora
       connection = Faraday.new(host, options)
       connection.basic_auth(user, password)
       connection
+    end
+
+    def validate_options
+      unless host.downcase.end_with?("/rest")
+        ActiveFedora::Base.logger.warn "Fedora URL (#{host}) does not end with /rest. This could be a problem. Check your fedora.yml config"
+      end
     end
   end
 end
